requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
mysql-connector-python==8.0.26
protobuf==4.23.1
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.4.0
